Add: API.

+++ routes/api.php
@@ -17,37 +17,37 @@ Route::prefix('scim/v2')
 ->middleware('requestlog')
 ->group(function (\Illuminate\Routing\Router $router) {
     // Users
-    $router->get('Users', 'AdminUserProvisionController@index')
+    $router->get('{tenant_id}/Users', 'AdminUserProvisionController@index')
     ->name('api.user.index');
     
-    $router->get('Users/{scim_id}', 'AdminUserProvisionController@show')
+    $router->get('{tenant_id}/Users/{scim_id}', 'AdminUserProvisionController@show')
     ->name('api.user.get');
     
-    $router->post('Users', 'AdminUserProvisionController@store')
+    $router->post('{tenant_id}/Users', 'AdminUserProvisionController@store')
     ->name('api.user.create');
     
-    $router->put('Users/{scim_id}', 'AdminUserProvisionController@replace')
+    $router->put('{tenant_id}/Users/{scim_id}', 'AdminUserProvisionController@replace')
     ->name('api.user.replace');
     
-    $router->patch('Users/{scim_id}', 'AdminUserProvisionController@update')
+    $router->patch('{tenant_id}/Users/{scim_id}', 'AdminUserProvisionController@update')
     ->name('api.user.update');
     
-    $router->delete('Users/{scim_id}', 'AdminUserProvisionController@delete')
+    $router->delete('{tenant_id}/Users/{scim_id}', 'AdminUserProvisionController@delete')
     ->name('api.user.delete');
     
     // Group
-    $router->get('Groups', 'AdminGroupProvisionController@index')
+    $router->get('{tenant_id}/Groups', 'AdminGroupProvisionController@index')
     ->name('api.group.index');
     
-    $router->post('Groups', 'AdminGroupProvisionController@store')
+    $router->post('{tenant_id}/Groups', 'AdminGroupProvisionController@store')
     ->name('api.group.create');
     
-    $router->get('Groups/{scim_id}', 'AdminGroupProvisionController@show')
+    $router->get('{tenant_id}/Groups/{scim_id}', 'AdminGroupProvisionController@show')
     ->name('api.group.show');
     
-    $router->patch('Groups/{scim_id}', 'AdminGroupProvisionController@update')
+    $router->patch('{tenant_id}/Groups/{scim_id}', 'AdminGroupProvisionController@update')
     ->name('api.group.update');
     
-    $router->delete('Groups/{scim_id}', 'AdminGroupProvisionController@delete')
+    $router->delete('{tenant_id}/Groups/{scim_id}', 'AdminGroupProvisionController@delete')
     ->name('api.group.delete');
 });
